Botmaster/01 - CleanAirMarketingCrew/requirements.txt
crewai>=0.11.0
langchain>=0.1.0
langchain-openai>=0.0.2
python-dotenv>=1.0.0
httpx>=0.24.0
litellm>=1.0.0
duckduckgo-search>=3.9.0
pyyaml>=6.0.1
pydantic>=2.0.0 